Reports were sometimes having very small rounding errors on very large reports. Changing our calculations to use type decimal instead of double to prevent this.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@8787 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDental/Forms Reports/FormQuery.cs
@@ -1061,7 +1061,7 @@ namespace OpenDental{
 			}//end for j cols
 			if(reportIn!=null){
 				for(int k=0;k<colTotals.Length;k++){
-					reportIn.ColTotal[k]=PIn.Double(colTotals[k].ToString("n"));
+					reportIn.ColTotal[k]=PIn.Decimal(colTotals[k].ToString("n"));
 				}
 			}
 			return tableOut;
